Fix explain

See merge request db/PGSpider!24

+++ contrib/pgspider_core_fdw/pgspider_core_fdw.c
@@ -222,6 +222,13 @@ enum SpdServerstatus
 	ServerStatusDead,
 };
 
+/* For EXPLAIN */
+static const char *SpdServerstatusStr[] = {
+	"Alive",
+	"Not specified by UNDER",
+	"Dead"
+};
+
 typedef struct Mappingcell
 {
 	/*
@@ -255,6 +262,7 @@ typedef struct ChildInfo
 	RelOptInfo *grouped_rel_local;
 	Agg		   *pAgg;			/* "Aggref" for Disable of aggregation push
 								 * down servers */
+	int			index_threadinfo;	/* index for ForeignScanThreadInfo array */
 }			ChildInfo;
 
 /*
@@ -296,6 +304,7 @@ typedef struct SpdFdwPrivate
 	bool		is_drop_temp_table; /* drop temp table flag in aggregation */
 	int			temp_num_cols;	/* number of columns of temp table */
 	char	   *temp_table_name;	/* name of temp table */
+	bool		is_explain;		/* explain or not */
 }			SpdFdwPrivate;
 
 /* postgresql.conf paramater */
@@ -2424,9 +2433,6 @@ spd_ExplainForeignScan(ForeignScanState *node,
 
 		fs = GetForeignServer(childinfo[i].server_oid);
 		fdwroutine = GetFdwRoutineByServerId(childinfo[i].server_oid);
-		/* skip to can not access child table at spd_GetForeignRelSize. */
-		if (childinfo[i].child_node_status != ServerStatusAlive)
-			continue;
 
 		if (fdwroutine->ExplainForeignScan == NULL)
 			continue;
@@ -2434,20 +2440,24 @@ spd_ExplainForeignScan(ForeignScanState *node,
 		/* create node info */
 		PG_TRY();
 		{
-			fsplan->fdw_private = ((ForeignScan *) childinfo[i].plan)->fdw_private;
+			int			idx;
 
-			/* TODO : fix */
-			fdwroutine->ExplainForeignScan(node, es);
+			ExplainPropertyText(psprintf(" * Node: %s / Status", fs->servername),
+								SpdServerstatusStr[childinfo[i].child_node_status], es);
+
+			if (childinfo[i].child_node_status != ServerStatusAlive)
+				continue;
+
 			if (es->verbose)
 			{
-				char	   *buf = "NodeName";
 
 				if (fdw_private->agg_query)
 				{
-					buf = psprintf("Agg push-down: %s / NodeName", childinfo[i].aggpath ? "no" : "yes");
+					ExplainPropertyText("Agg push-down", childinfo[i].aggpath ? "no" : "yes", es);
 				}
-				ExplainPropertyText(buf, fs->servername, es);
 			}
+			idx = childinfo[i].index_threadinfo;
+			fdwroutine->ExplainForeignScan(((ForeignScanThreadInfo *) node->spd_fsstate)[idx].fsstate, es);
 
 		}
 		PG_CATCH();
@@ -3068,13 +3078,20 @@ spd_BeginForeignScan(ForeignScanState *node, int eflags)
 	 */
 	hash_register_reset_callback(node->ss.ps.state->es_query_cxt);
 
-	if (eflags & EXEC_FLAG_EXPLAIN_ONLY)
-		return;
 
 	oldcontext = MemoryContextSwitchTo(TopTransactionContext);
 	node->spd_fsstate = NULL;
 	fdw_private = (SpdFdwPrivate *)
 		((Value *) list_nth(fsplan->fdw_private, FdwScanPrivateSelectSql))->val.ival;
+
+	/*
+	 * Not return from this function unlike usual fdw BeginForeignScan
+	 * implementation because we need to create ForeignScanState for child
+	 * fdws. It is assigned to fssThrdInfo[node_incr].fsstate.
+	 */
+	if (eflags & EXEC_FLAG_EXPLAIN_ONLY)
+		fdw_private->is_explain = true;
+
 	/* Type of Query to be used for computing intermediate results */
 #ifdef GETPROGRESS_ENABLED
 	if (fdw_private->agg_query)
@@ -3249,8 +3266,10 @@ spd_BeginForeignScan(ForeignScanState *node, int eflags)
 		fssThrdInfo[node_incr].iFlag = true;
 		fssThrdInfo[node_incr].EndFlag = false;
 		fssThrdInfo[node_incr].tuple = NULL;
-		/* We set index of child info, not set node_incr */
+
+		/* We save correspondence between fssThrdInfo and childinfo */
 		fssThrdInfo[node_incr].childInfoIndex = i;
+		childinfo[i].index_threadinfo = node_incr;
 
 		fssThrdInfo[node_incr].serverId = server_oid;
 		fssThrdInfo[node_incr].fdwroutine = GetFdwRoutineByServerId(server_oid);
@@ -3259,19 +3278,42 @@ spd_BeginForeignScan(ForeignScanState *node, int eflags)
 			ResourceOwnerCreate(CurrentResourceOwner, "thread resource owner");
 
 		fssThrdInfo[node_incr].private = fdw_private;
+
+		/*
+		 * For explain case, call BeginForeignScan because some
+		 * fdws(ex:mysql_fdw) requires BeginForeignScan is already called when
+		 * ExplainForeignScan is called . For non explain case, child threads
+		 * call BeginForeignScan
+		 */
+		if (eflags & EXEC_FLAG_EXPLAIN_ONLY)
+			fssThrdInfo[node_incr].fdwroutine->BeginForeignScan(fssThrdInfo[node_incr].fsstate,
+																eflags);
+		node_incr++;
+	}
+
+	fdw_private->nThreads = node_incr;
+
+	/* Skip thread creation in explain case */
+	if (eflags & EXEC_FLAG_EXPLAIN_ONLY)
+	{
+		MemoryContextSwitchTo(oldcontext);
+		return;
+	}
+
+	for (i = 0; i < fdw_private->nThreads; i++)
+	{
 		thread_create_err =
-			pthread_create(&fdw_private->foreign_scan_threads[node_incr],
+			pthread_create(&fdw_private->foreign_scan_threads[i],
 						   NULL,
 						   &spd_ForeignScan_thread,
-						   (void *) &fssThrdInfo[node_incr]);
+						   (void *) &fssThrdInfo[i]);
 		if (thread_create_err != 0)
 		{
 			ereport(ERROR, (errmsg("Cannot create thread! error=%d",
 								   thread_create_err)));
 		}
-		node_incr++;
 	}
-	fdw_private->nThreads = node_incr;
+
 
 	/* Wait for state change */
 	for (node_incr = 0; node_incr < fdw_private->nThreads; node_incr++)
@@ -4182,22 +4224,25 @@ spd_EndForeignScan(ForeignScanState *node)
 	if (isPrintError)
 		spd_PrintError(fdw_private->node_num, fdw_private->childinfo);
 
-	if (fdw_private->is_drop_temp_table == FALSE)
+	if (!fdw_private->is_explain)
 	{
-		spd_spi_ddl_table(psprintf("DROP TABLE IF EXISTS %s", fdw_private->temp_table_name));
-	}
-	for (node_incr = 0; node_incr < fdw_private->nThreads; node_incr++)
-	{
-		fssThrdInfo[node_incr].EndFlag = true;
+		if (fdw_private->is_drop_temp_table == FALSE)
+		{
+			spd_spi_ddl_table(psprintf("DROP TABLE IF EXISTS %s", fdw_private->temp_table_name));
+		}
+		for (node_incr = 0; node_incr < fdw_private->nThreads; node_incr++)
+		{
+			fssThrdInfo[node_incr].EndFlag = true;
+			/* Cleanup the thread-local structures */
+			rtn = pthread_join(fdw_private->foreign_scan_threads[node_incr], NULL);
+			if (rtn != 0)
+				elog(WARNING, "error is occurred, pthread_join fail in EndForeignScan. ");
+		}
 	}
 
 	/* wait until all the remote connections get closed. */
 	for (node_incr = 0; node_incr < fdw_private->nThreads; node_incr++)
 	{
-		/* Cleanup the thread-local structures */
-		rtn = pthread_join(fdw_private->foreign_scan_threads[node_incr], NULL);
-		if (rtn != 0)
-			elog(WARNING, "error is occurred, pthread_join fail in EndForeignScan. ");
 		if (fssThrdInfo[node_incr].fsstate->ss.ss_currentRelation)
 			RelationClose(fssThrdInfo[node_incr].fsstate->ss.ss_currentRelation);
 		pfree(fssThrdInfo[node_incr].fsstate);
